Rename WholesaleInbox related types
